Implementando o serviço de mensageria e finalizando o endpoint mensagem_candidato

+++ src/main/java/com/example/app/model/entities/Alerta.java
@@ -11,17 +11,17 @@ public class Alerta {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	
-	private Long candidaturaId;
+	private Integer candidaturaId;
 	private String mensagem;
 	
 	public Alerta() {
 	}
 
-	public Long getCandidaturaId() {
+	public Integer getCandidaturaId() {
 		return candidaturaId;
 	}
 
-	public void setCandidaturaId(Long candidaturaId) {
+	public void setCandidaturaId(Integer candidaturaId) {
 		this.candidaturaId = candidaturaId;
 	}
